move default bot image to users package
